Split ActionChooser.py into ActionChooserParts/
